python/jdbc/requirements.txt
JayDeBeApi>=1.2.3
JPype1>=1.4.0
pandas>=1.5.0
